UPSTREAM: mb/google/sarien: Support multiple touchscreen at same address
The Sarien board may have different touchscreen devices that use the
same I2C slave address but have different requirements such as needing
a special driver or ACPI configuration.

In order to support this the devicetree may be configured with multiple
devices at the same address and at boot time the unused devices will be
disabled.

Because there is no GPIO for selecting the device that is present it can
instead be selected with Kconfig, or by setting a VPD key to the HID of
the touchscreen device that is present.  The default for Sarien devices
is to not enable a touchscreen for the OS.

The touchscreen selection is currently limited to the Sarien variant but
this also adds the touchscreen HID for Arcada to Kconfig so it would not
complain about the key not being set.

BUG=b:122019253
TEST=This was tested on a Sarien board by adding a second entry to the
devicetree at the same address.  Without this change the SSDT is not
loaded by the kernel because of the address conflict.  After this change
no touchscreen is enabled by default, but one can be selected with
Kconfig or by setting the 'touchscreen_hid' VPD key.

Change-Id: I96c98d1f789d4b07e209677886b4c2706ef7baf9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f1690f0ec166c214a2c42acf2d410ea7f35a77f0
Original-Change-Id: I4da12b1de0c551bcd89325fe0d8c66c6ffeb7afc
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/31295
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1470588
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>