UPSTREAM: ec/lenovo/h8/acpi: Add alternative Fn-F2 and Fn-F3 layout
thinkpad_acpi maps the battery hotkey (KEY_BATTERY) on scancode 0x01 and
the lock hotkey (KEY_COFFEE) on scancode 0x02.

On the Thinkpad X1 Carbon (and possibly others), the hotkeys for Fn-F2
and Fn-F3 are different from the default one so a new layout has to be
defined.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia1da626e8d1140470a1de7345a542d5c813bcfc5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fab9ae8167cc1f830452bb3da0157736fb7d4245
Original-Change-Id: Ib2d96be1a7815d7d03e6e8c6d300fd671c8598ca
Original-Signed-off-by: Nicola Corna <nicola@corna.info>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31470
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2062809
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>