UPSTREAM: drivers/spi/spi_flash.c: Avoid static scan false positive
Static scan-build indicates a possible invalid return from function
spi_flash_cmd_erase(). The root cause is because the scan believes it's
possible for offset to be above the end address in the first pass, thus
not setting a value for variable ret. Assign initial value of -1 to
variable ret to make checker happy.

BUG=b:112253891
TEST=build grunt

Change-Id: I3d4a700fa021767b752cbadd405623c4bc01a6b7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 22ba205f7024fd04436c1ad415c26d6961919ef4
Original-Change-Id: If548728ff90b755c69143eabff6aeff01e8fd483
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/c/31473
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1478166
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>