UPSTREAM: soc/intel/cannonlake: SoC specific microcode update check
For CFL and WHL, Microcode is being loaded from FIT. Both
supports the PRMRR/SGX feature. If This is supported the FIT
microcode load will set the msr (0x08b) with the Patch id one
less than the id in the microcode binary. This results in
Microcode getting reloaded again in bootblock and ramstage.
Avoid the microcode reload by checking for PRMRR support.
CFL and WHL CPU die are based on KBL CPU so we need to have
this check, where CNL CPU die is not based on KBL CPU so
skip this check for CNL.

BUG=b:124126405
Signed-off-by: Ronak Kanabar <ronak.kanabar@intel.com>

Change-Id: Id796e92d76e158f67db0f785dccabfacd6c8fec6
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id:
Original-Change-Id: Ie1ce5881358bae95f60281bebd92175256d677fd
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Original-Commit-Id: 69a95653399657bf6b2491c259479c954fee2509
Original-Original-Change-Id: I3311a7413d27044f9c819179e5b0cb9a67b46955
Original-Original-Reviewed-on: https://review.coreboot.org/c/31492
Original-Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Original-Reviewed-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1485877
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>