UPSTREAM: rockchip/rk3399: Fix BL31 bootmem regions
The BL31 on RK3399 is split into multiple segments... the majority goes
into DRAM, but small parts must be put into SRAM and PMUSRAM. With
CB:31123 only the DRAM part was added to memlayout, so the SRAM parts
will not be correctly marked in bootmem and BL31 loading fails the
selfload check. This patch adds the remaining regions to fix the
problem.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia3c125b0514e43be61cd0e0c9ae60926d39c4dbf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id:
Original-Change-Id: I148503e9f0af12b307961e041f6b73fa0a047986
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Original-Commit-Id: 314b5c370b4655bc701985ddf1d1d478067e7baa
Original-Original-Change-Id: Ia0597216c08512c47361a1dc0beb34d022a8994f
Original-Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Original-Reviewed-on: https://review.coreboot.org/c/31538
Original-Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Original-Reviewed-by: Ting Shen <phoenixshen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1485884
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>