UPSTREAM: mb/lenovo/t430s: Open PCIe port #5 for Thunderbolt controller
Some T430s variants have a Thunderbolt controller wired to PCIe port

The controller hotplugs itself to the chipset when a downstream device
is hotplugged into it, so the hotplug capability should be enabled on
PCIe port #5.

TODO: find the correct gpio pin to detect the Thunderbolt controller
at runtime.

There are 3 variants of mainboard for Thinkpad T430s: Basic type
(Wistron LSN-4 11263-1), Boards with an additional discreet GPU,
Boards with an additional TB controller (Wistron LSN-4 11271-1),
each of which has a different schematic.

The gpio27 on the last type is set as set as GPIO-INPUT, compared
with GPIO-OUTPUT-HIGH on the basic type boards.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5595917a998591cf532670e0c8693675712614c0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0ffdeef681f5f17b1aa66d9b99f5501089ea7b1a
Original-Change-Id: I61f41db100f398069e50e2da8a378b3a8d1c84bf
Original-Signed-off-by: Bill XIE <persmule@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31631
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1809149
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>