UPSTREAM: driver/intel/ish: add ish chip driver support
We want to be able to specify the firmware variant suffix
in the devicetree.cb configuration for particular firmware
builds. This driver allows us to specify the firmware_variant
property in the device tree and have it populate a _DST table
in the SSDT ACPI table for the ISH device, thus making the
suffix available to the kernel (See crrev.com/c/1433482 for kernel
change that uses the value)

BUG=b:122722008
TEST=decompile DDST table and verify that new firmware-variant value
is present. Also verfied that kernel can access this new field using
the shim loader kernel CLs

Change-Id: I28e646f2fc021e58f7ded515f23a8f39c723ca47
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6bdfc8027bb68bcecd7c31069065c1e7c8ef3598
Original-Change-Id: Id8be986185282521aee574027503eaf8968e1508
Original-Signed-off-by: Jett Rink <jettrink@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31682
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1506741
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>