UPSTREAM: lint/kconfig: Update to support new CONFIG() macro
This patch updates the Kconfig linter to support the new CONFIG() macro
in the same manner that IS_ENABLED() was previously supported. It will
be flagged when it is used on non-bool Kconfigs or used with #ifdef, and
it is supported for checking used Kconfigs. Remaining uses of
IS_ENABLED() are flagged with a deprecation warning.

BUG=none
BRANCH=none
TEST=none

Change-Id: I521ff92d45e751bea7433e2529be6fe78e851279
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: ef7a3267870f126cc2f815812cfe54500853d2b8
Original-Change-Id: I171ea8bc8e2d22abab7fc4d87ff4cf8aad21084f
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/31776
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1518308